refactor(api): 抽取共享响应类型定义
