+++ src/api/types.ts
@@ -1,0 +1,7 @@
+export interface ApiResponse<T = unknown> {
+  result: boolean;
+  data?: T;
+  err?: string;
+  message?: string;
+  [key: string]: any;
+}
